+++ beetsplug/bandcamp/_helpers.py
@@ -84,6 +84,7 @@ rm_strings = [
     r"^[EL]P( \d+)?",
     r"^Vol(ume)?\W*(?!.*\)$)\d+",
     r"\((digital )?album\)",
+    r"\(single\)",
     r"^va|va$|vinyl(-only)?|compiled by.*",
     r"free download|\([^()]*free(?!.*mix)[^()]*\)",
 ]